set up so that only a characters owner can delete a character

+++ backend/server/v2/delete/deleteV2CharacterController.ts
@@ -1,4 +1,5 @@
 import { checkForContentTypeBeforeSending } from '../../controllers/common/sendingFunctions'
+import query from '../../db/database'
 import { Response, Request } from '../../interfaces/apiInterfaces'
 import deleteFromCharacterPages from './utilities/deleteFromCharacterPages'
 import deleteFromOwner from './utilities/deleteFromOwner'
@@ -10,17 +11,23 @@ export interface ViewRequest extends Request {
     }
 }
 
+const checkOwnerSQL = `select ownerID = $1 as isOwner from v2CharacterOwner
+where characterID = $2`
+
 export async function deleteV2Character(request: ViewRequest, response: Response) {
     const characterID = +request.params.characterID
 
-    // TODO
-    // check if owner
+    const [{isowner: isOwner}] = await query(checkOwnerSQL, [request.user?.id, characterID])
 
-    await Promise.all([
-        deleteFromOwner(characterID),
-        deletePages(characterID),
-        deleteFromCharacterPages(characterID)
-    ])
-
-    checkForContentTypeBeforeSending(response, { message: 'Deleted' })
+    if (isOwner) {
+        await Promise.all([
+            deleteFromOwner(characterID),
+            deletePages(characterID),
+            deleteFromCharacterPages(characterID)
+        ])
+    
+        checkForContentTypeBeforeSending(response, { message: 'Deleted' })
+    } else {
+        checkForContentTypeBeforeSending(response, { message: 'You Don\'t Own This Character' })
+    }
 }
